Firebase build
For sucessfull builds you will need to create realtime database in Firebase and download google-services.json and put it inside of app or src (dont remmeber it, where it has to be, I had it in both folders) folder.

Readme will have example structure how should look your Firebase database. (The app is in slovak language, so you might do some tinkering if you will want to rename variables or translations, since it was build just for me and now decided to make it public).

The old offline version is in archived branch.